rentquest/api/views.py
from rest_framework import viewsets
from .models import User, Property
from .serializers import UserSerializer, PropertySerializer


class UserViewSet(viewsets.ModelViewSet):
    queryset = User.objects.all()
    serializer_class = UserSerializer


class PropertyViewSet(viewsets.ModelViewSet):
    queryset = Property.objects.all()
    serializer_class = PropertySerializer
